requirements.txt
setuptools
black==19.10b0
Django==3.0.6
djangorestframework==3.11.0
flake8==3.8.2
psycopg2==2.8.5
python-crontab==2.5.1
